// sourcecode/BTrak.Packages/snova-leave-management/leave-management-module/leave-management-components/src/lib/leavemanagement/models/leave-model.ts
export class LeaveModel {
    leaveApplicationId: string;
    leaveApplicationIds: string;
    userId: string;
    leaveReason: string;
    leaveTypeId: string;
    leaveStatusId: string;
    overallLeaveStatusId: string;
    leaveDateFrom: Date;
    leaveDateTo: Date;
    isDeleted: boolean;
    isArchived: boolean;
    fromLeaveSessionId: string;
    toLeaveSessionId: string;
    date: Date;
    pageSize: number;
    pageNumber: number;
    sortBy: string;
    sortDirectionAsc: boolean;
    leaveStatusName: string;
    timeStamp: any;
    isApproved: boolean;
    isWaitingForApproval: boolean;
    totalCount: number;
    isToReApply: boolean;
    branchId: string;
    isAdmin: boolean;
    employeeName: string;
    isRejected: boolean;
    leaveStatusColour: string;
    isAdminApprove: boolean;
    start: Date;
    end: Date;
    id: string;
    isCalendarView: boolean;
    title: string;
    dataItem: any;
    profileImage: string;
    entityId: string;
    leaveTypeColor: string;
}